chore: Added comments to explain `shutdown`.
- The use of shutdown over close is a rather nuanced topic
for UDP servers. For the time being I am using shutdown as a
way to guarantee that the file descriptor isn't reused until
the last reference to the file descriptor is destroyed. I would
like to avoid the need to add an explicit file-descriptor `remove`
method to the asynchronous event-loop as that will create new
synchronization problems.